understanding and navigation in self-driving cars.
Dear Students welcome to new project. Analyzing urban environments for various applications such as traffic management and infrastructure planning.

### Data Format: include
-Images: The dataset contains RGB images extracted from video sequences.
-Labels: Corresponding ground truth annotations in the form of labeled images where each pixel's value represents a class.

Overview:

The CamVid (Cambridge-driving Labeled Video Database) dataset is a comprehensive collection of videos and corresponding annotations designed for the purpose of semantic segmentation and scene understanding. The dataset was collected by the Machine Intelligence Laboratory at the University of Cambridge. It is widely used for training and evaluating machine learning models, particularly in the field of autonomous driving and computer vision.